Restructured bme280 library
