+++ stm32/lib/bme280/include/bme280_common.h
@@ -1,0 +1,140 @@
+/**\
+ * Copyright (c) 2020 Bosch Sensortec GmbH. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ **/
+
+#ifndef __BME280_COMMON_H__
+#define __BME280_COMMON_H__
+
+/*! CPP guard */
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include <stdio.h>
+#include "bme280.h"
+
+/***************************************************************************/
+
+/*!                 User function prototypes
+ ****************************************************************************/
+
+/*!
+ *  @brief Function for reading the sensor's registers through SPI bus.
+ *
+ * @param[in] reg_addr       : Register address from which data is read.
+ * @param[out] reg_data      : Pointer to data buffer where read data is stored.
+ * @param[in] length         : Number of bytes of data to be read.
+ * @param[in, out] intf_ptr  : Void pointer that can enable the linking of descriptors
+ *                             for interface related call backs.
+ *
+ *  @return Status of execution
+ *
+ *  @retval BME280_INTF_RET_SUCCESS -> Success.
+ *  @retval != BME280_INTF_RET_SUCCESS -> Failure.
+ *
+ */
+BME280_INTF_RET_TYPE bme280_spi_read(uint8_t reg_addr, uint8_t *reg_data, uint32_t length, void *intf_ptr);
+
+/*!
+ *  @brief Function for reading the sensor's registers through I2C bus.
+ *
+ * @param[in] reg_addr       : Register address from which data is read.
+ * @param[out] reg_data      : Pointer to data buffer where read data is stored.
+ * @param[in] length         : Number of bytes of data to be read.
+ * @param[in, out] intf_ptr  : Void pointer that can enable the linking of descriptors
+ *                             for interface related call backs.
+ *
+ *  @return Status of execution
+ *
+ *  @retval BME280_INTF_RET_SUCCESS -> Success.
+ *  @retval != BME280_INTF_RET_SUCCESS -> Failure.
+ *
+ */
+BME280_INTF_RET_TYPE bme280_i2c_read(uint8_t reg_addr, uint8_t *reg_data, uint32_t length, void *intf_ptr);
+
+/*!
+ *  @brief Function for writing the sensor's registers through SPI bus.
+ *
+ * @param[in] reg_addr      : Register address to which the data is written.
+ * @param[in] reg_data      : Pointer to data buffer in which data to be written
+ *                            is stored.
+ * @param[in] length        : Number of bytes of data to be written.
+ * @param[in, out] intf_ptr : Void pointer that can enable the linking of descriptors
+ *                            for interface related call backs
+ *
+ *  @return Status of execution
+ *
+ *  @retval BME280_INTF_RET_SUCCESS -> Success.
+ *  @retval  != BME280_INTF_RET_SUCCESS -> Failure.
+ *
+ */
+BME280_INTF_RET_TYPE bme280_spi_write(uint8_t reg_addr, const uint8_t *reg_data, uint32_t length, void *intf_ptr);
+
+/*!
+ *  @brief Function for writing the sensor's registers through I2C bus.
+ *
+ * @param[in] reg_addr      : Register address to which the data is written.
+ * @param[in] reg_data      : Pointer to data buffer in which data to be written
+ *                            is stored.
+ * @param[in] length        : Number of bytes of data to be written.
+ * @param[in, out] intf_ptr : Void pointer that can enable the linking of descriptors
+ *                            for interface related call backs
+ *
+ *  @return Status of execution
+ *
+ *  @retval BME280_INTF_RET_SUCCESS -> Success.
+ *  @retval != BME280_INTF_RET_SUCCESS -> Failure.
+ *
+ */
+BME280_INTF_RET_TYPE bme280_i2c_write(uint8_t reg_addr, const uint8_t *reg_data, uint32_t length, void *intf_ptr);
+
+/*!
+ *  @brief This function provides the delay for required time (Microsecond) as per the input provided in some of the
+ *  APIs.
+ *
+ *  @param[in] period_us  : The required wait time in microsecond.
+ *  @param[in] intf_ptr   : Interface pointer
+ *
+ *  @return void.
+ */
+void bme280_delay_us(uint32_t period_us, void *intf_ptr);
+
+/*!
+ *  @brief This function is to select the interface between SPI and I2C.
+ *
+ *  @param[in] dev    : Structure instance of bme280_dev
+ *  @param[in] intf   : Interface selection parameter
+ *                          For I2C : BME280_I2C_INTF
+ *                          For SPI : BME280_SPI_INTF
+ *
+ *  @return Status of execution
+ *  @retval 0 -> Success
+ *  @retval < 0 -> Failure
+ */
+int8_t bme280_interface_selection(struct bme280_dev *dev, uint8_t intf);
+
+/*!
+ *  @brief This API is used to print the execution status.
+ *
+ *  @param[in] api_name : Name of the API whose execution status has to be printed.
+ *  @param[in] rslt     : Error code returned by the API whose execution status has to be printed.
+ *
+ *  @return void.
+ */
+void bme280_error_codes_print_result(const char api_name[], int8_t rslt);
+
+/*!
+ * @brief This function deinitializes coines platform
+ *
+ *  @return void.
+ *
+ */
+void bme280_coines_deinit(void);
+
+#ifdef __cplusplus
+}
+#endif /* End of CPP guard */
+
+#endif /* __BME280_COMMON_H__ */
