+++ stm32/lib/bme280/include/bme280_defs.h
